--- packages/log/ts/memory/logEntriesToString.ts
@@ -1,7 +1,0 @@
-import { plainLogFormatter } from '../formatter/index.js'
-import { LogEntry } from '../types.js'
-import { toInspectLogEntry } from './toInspectStringForObject.js'
-
-export function logEntriesToString(logs: LogEntry[]) {
-  return logs.map(toInspectLogEntry).map(plainLogFormatter).join('\n')
-}
