--- packages/log/ts/memory/toInspectStringForObject.browser.ts
@@ -1,3 +1,0 @@
-import { LogEntry } from '../LogEntry.js'
-
-export const toInspectLogEntry = (e: LogEntry) => e
