+++ packages/log/ts/platform/index.browser.ts
@@ -1,3 +1,7 @@
+import { LogEntry } from '../LogEntry.js'
+
+export const toInspectLogEntry = (e: LogEntry) => e
 export function isConsoleDebugAvailable() {
   return true
 }
+
