+++ packages/log/ts/platform/index.ts
@@ -1,3 +1,6 @@
+import { inspect } from 'util'
+import { LogEntry } from '../LogEntry.js'
+
 export function semverGt(versionString: string, version: [number, number, number]) {
   const actual = versionString.split('.').reverse().reduce((p, v, i) => {
     p += Number.parseInt(v, 10) * Math.pow(100, i)
@@ -13,3 +16,7 @@ export function isConsoleDebugAvailable() {
   const versionString = process.version.startsWith('v') ? process.version.slice(1) : process.version
   return semverGt(versionString, [9, 3, 0])
 }
+
+export const toInspectLogEntry = (e: LogEntry) => (
+  e.args = e.args.map(value => typeof value === 'object' && value !== null ? inspect(value) : value),
+  e)
